Переделать превью сторибука
